requirements.txt
keras
git+https://www.github.com/keras-team/keras-contrib.git
matplotlib
numpy
scipy
pillow
scikit-image
scikit-learn
opencv-python
pandas
tqdm